+++ src/ballblast/model/gameobjects/Bullet.java
@@ -9,11 +9,11 @@ import com.google.common.base.Objects;
  *
  */
 public class Bullet extends AbstractGameObject {
-    private static final int DEFAULT_WIDTH = 12;
-    private static final int DEFAULT_HEIGHT = 15;
+    private static final double DEFAULT_WIDTH = 12;
+    private static final double DEFAULT_HEIGHT = 15;
 
-    private final int height;
-    private final int width;
+    private final double height;
+    private final double width;
     /**
      * Create a {@link Bullet} instance.
      */
@@ -24,12 +24,12 @@ public class Bullet extends AbstractGameObject {
     }
 
     @Override
-    public final int getHeight() {
+    public final double getHeight() {
         return this.height;
     }
 
     @Override
-    public final int getWidth() {
+    public final double getWidth() {
         return this.width;
     }
